Reorganize custom model API into unified custom_models/ directory
This commit reorganizes the custom model integration code for better
discoverability and cleaner API surface:

**vLLM core changes:**
- Move trainable attention files to vllm/model_executor/custom_models/
  - trainable_attention.py (TrainableFlashAttention)
  - trainable_mla_attention.py (TrainableMLA)
  - attention_replacement.py (replace_with_trainable_attention)
  - custom_model_wrapper.py (VLLMModelForCausalLM base class)

- Update vllm/model_executor/custom_models/__init__.py to export all API
- Update all import paths throughout codebase:
  - Example files (deepseek_v3, qwen3, megatron example)
  - Test files (test_generic_models.py)
  - Documentation (basic.md)
  - Internal references (layers/__init__.py)

**Example reorganization:**
- Create examples/custom_models/ directory
- Move deepseek_v3_torchtitan.py and qwen3_torchtitan.py to examples/custom_models/
- Add examples/custom_models/README.md with usage instructions

**Test reorganization:**
- Create tests/custom_models/ directory
- Move test_deepseek_v3_tp.py to tests/custom_models/
- Convert to proper pytest test with fixtures and assertions

**Documentation:**
- Add docs/contributing/model/custom.md with comprehensive guide
  - Quick start example
  - Core components reference
  - Advanced topics (TP integration, multi-GPU)
  - Complete examples and troubleshooting

**Benefits:**
- Single unified import: from vllm.model_executor.custom_models import ...
- Clear separation of custom model code from vLLM internals
- Better discoverability for users
- Proper test organization
- Comprehensive documentation

All imports updated, tests pass, documentation complete.
